Some refactoring.

+++ Software/org.whitefossa.yiffhl/org.whitefossa.yiffhl/org.whitefossa.yiffhl/Abstractions/DTOs/FoxIdentificationData.cs
@@ -10,16 +10,16 @@ namespace org.whitefossa.yiffhl.Abstractions.DTOs
         /// <summary>
         /// Fox hardware revision
         /// </summary>
-        public UInt16 HardwareRevision { get; set; }
+        public uint HardwareRevision { get; set; }
 
         /// <summary>
         /// Firmware version
         /// </summary>
-        public UInt16 FirmwareVersion { get; set; }
+        public uint FirmwareVersion { get; set; }
 
         /// <summary>
         /// Serial number
         /// </summary>
-        public UInt32 SerialNumber { get; set; }
+        public uint SerialNumber { get; set; }
     }
 }
